Fix use of secrets on WP Cloud site (#1909)
## Motivation for the change, related issues

A previous change to how we serve PHP files on the website broke use of
secrets stored with WP Cloud. This PR fixes that.

## Implementation details

Because we are using WP cloud persistent data storage early in the
request lifecycle (when custom-redirects.php is executed), certain
environmental dependencies of persistent storage are not yet defined.
This PR forces the definition of those dependencies to give us access to
secrets from persistent storage.

## Testing Instructions (or ideally a Blueprint)

Manually tested on Playground website.