Fix progress reporting during Playground load
Prior to this change, the service worker was not making the response
available to consumers until the response was placed in the offline
cache. It looks like the entire response had to be buffered before
consumers had a chance to observe response progress, and at that
point, there was no progress to observe because the download was
complete.

This updates service worker request caching so that responses are
made available downstream before caching is complete.