[CORS Proxy] Rate-limits IPv6 requests based on /64 subnets, not specific addresses (#1923)
Rate-limits the CORS proxy requests based on the first 64 bits of an
IPv6 address, not all 128 bits. This prevents a person with an entire
/64 subnet from exhausting the storage or getting more than their fair
share of the tokens. This only applies to IPv6. For IPv4 addresses, all
64 bits are still considered.

 ## Implementation

Converts a string-based IP address into a binary string, then zeros the
first 64 bits in that string and re-encodes it as a human-readable IP
string.

 ## Testing instructions

* Deploy and attack the proxy :-)
* Run the unit tests

```php
php ./packages/playground/website-deployment/tests.php
```

cc @brandonpayton for reviews

---------

Co-authored-by: Brandon Payton <brandon@happycode.net>