Fix changelog workflow and backfill changelog entries from recent versions (#83)
## Motivation for the change, related issues

Our changelog workflow is currently broken because it is missing extra
GitHub token secrets. If we add these secrets, we will have to update
them occasionally. I think we may be able to do without and would like
to try that.

In addition, we need to backfill changelog entries while omitting links
to PRs (since we cannot link to our private PRs).

## Implementation details

This PR:
- Updates our changelog update workflow to persist credentials in the
local git config and attempts to follow the example of pushing a commit
after actions/checkout:

https://github.com/actions/checkout?tab=readme-ov-file#push-a-commit-using-the-built-in-token
- Omits PR links from new changelog entries unless they point to
github.com/WordPress/wordpress-playground
- Backfills changelog entries for v1.0.25 to v1.0.29.

## Testing Instructions (or ideally a Blueprint)

- Temporarily disable protections on running the workflow
- Make the workflow target the PR branch
- Manually run the workflow and see if it commits a changelog update to
the PR branch
- If successful, remove the changelog commit from the PR branch
- Re-enable protections
- Re-enable checking out trunk only
- Merge