Use Vite conventions for resolving file path
This way, the path can be resolved both before and
after build. Our custom module loader that supports
running in Node without a build step supports
importing URLs via the "?url" suffix, and at build time,
Vite replaces the same "?url" import with the URL
of the import after build.

Before this change, we used __dirname to resolve
the file path. This broke running unbuilt modules in
Node.js because esm modules do not have access
to a __dirname global.