Follow symlinks in useHostFilesystem()
useHostFilesystem() mounts all top-level directories from the host filesystem in PHP.wasm's VFS.

This PR ensures it also mounts top-level symlinks to directories.

For example, on Mac, the top level /var directory is a symlink
to /private/var. If we don't follow the symlink, we won't mount
it and PHP will use a separate, VFS-scoped /var directory. This,
in turn, will break the following use-case:

 * PHP.wasm writes a PHP script to a temporary directory in
 * /private/var/folders/.../T/ in the host filesystem.
 * PHP.wasm calls proc_open() to execute the script in the host filesystem.

This PR solves it.