PHP CLI – Ensure the PHP_BINARY constant is set to the PHP-WASM binary
## Rationale

We want any `proc_open()` calls to use the PHP-WASM binary and
not the system PHP binary.

## How it works

The code below creates a temporary `php` executable in PATH,
which covers `proc_open( "php", ... )` calls.

Furthermore, when PHP detects the `php` executable in PATH, it
sets the PHP_BINARY constant to it.

## Remaining work

Test coverage