[PHP] Ensure PHP_BINARY is defined in CLI scripts
In CLI, PHP defines a constant called `PHP_BINARY`. Many CLI scripts use it for `proc_open(PHP_BINARY)`. This PR makes sure that constant is always defined when running PHP.wasm in `php.cli()` mode.

 ## Implementation

This PR ships a fallback `/internal/shared/bin/php` binary for PHP to find when there are no other matching executables in the path. The changes are in phpwasm-emscripten-library.js since every PHP build will always need that fallback.

PHP_BINARY is documented as the path of the currently running PHP binary, but in practice it's just the first executable `php` binary found in PATH.

 ## Testing instructions

I've added a CI test – confirm it passes.