[PHP] Refresh the latest PHP versions before recompiling
This PR refreshes the supported-php-versions.mjs file with the latest
PHP versions when you attempt to rebuild any PHP version (but only if
the list is older than 24 hours).

This will help Playground stay up to date with latest PHP point
releases. Right now it's a few point versions behind.

 ## Testing instructions

* Manually set `lastRefreshed` to a past date in `supported-php-versions`
* Run npm run recompile:php:node:asyncify:8.4
* Confirm the versions were refreshed before the build and that it
  builds 8.4.10 instead of 8.4.0