Patch the premature "request in progress" semaphore release
Guards against releasing the "request in progress" semaphore
too early in the php.runStream() call.

A single PHP runtime can only handle one request at a time.
The PHP class calls a `wasm_sapi_handle_request` C function that
initializes the PHP runtime and starts the request. That function is
asynchronous and may yield back to the event loop before the request
is fully handled, the exit code known, and the runtime is cleaned up
and prepared for another request.

The PHP class uses an async semaphore to protect against calling
`wasm_sapi_handle_request` again while a previous call is still
running.

However, PR 2266 [1] introduced a regression where the semaphore
was released too early. As a result, it opened the runtime to a
race condition where a subsequent runStream() call tried to run
PHP code on a runtime that was in a middle of handling a request.

This test ensures that two runStream() calls can be made without
crashing the runtime.

[1] https://github.com/WordPress/wordpress-playground/pull/2266