[PHP] Include CLI SAPI in the web PHP.wasm build
This PR ships the web wasm binaries with the `run_cli` C function that
enables calling php.cli(). Intertwining the CLI SAPI usage with the web
SAPI usage was impractical before #2559, but now we can easily call
`await php.cli(); await php.run()` on the same PHP instance.

Note this PR does not expose the `.cli()` method to the public yet. It
only ships the updated WASM binaries to unlock that.

 ## Testing instructions

 CI