[Website] Limit the callstack size in filesystem journal
Replaces recursion with a loop in the filesystem journal code to avoid
the following error when installing Woo on an OPFS site:

```
Uncaught (in promise) RangeError: Maximum call stack size exceeded
    at normalizeFilesystemOperations (fs-journal.ts:364:17)
    at normalizeFilesystemOperations (fs-journal.ts:473:11)
    ...
```

\## Implementation

`normalizeFilesystemOperations()` used to call itself in a way that
significantly inflated the callstack when either there was either a
lot of operations to normalize, or the normalized operations referenced
each other recursively.

This PR replaces the recursive call with a while(true) loop which
limits the callstack size and allows the normalization to finish.

\## Testing instructions

Create a new site, save it "to this browser", go to wp-admin, loop up
Woo, click install, confirm it worked.