[PHP] Move setImmediate polyfill to loadWebRuntime
Follows up on #3021

Moves the setImmediate polyfill into loadWebRuntime exported from
`@php-wasm/web` and makes it conditional on the presence of
`setImmediate` in `globalThis`. This way:

* All `@php-wasm/web` consumers will benefit from the fix, not just the
  Playground website.
* If another `setImmediate` polyfill is used (from outside of
  Emscripten), we will not overrite it.

 ## Testing instructions

This PR ships with an e2e test that confirms the memory is deallocated
after running a Blueprint. Confirm it passes in the CI.