Fix extension path resolution in split php-wasm packages
The getIntlExtensionPath() and getXdebugExtensionPath() functions
in each @php-wasm/node-* package were using currentDirPath (which
resolves to the src/ directory) to build paths to extension files.

However, the asyncify/ and jspi/ directories containing the .so files
are siblings of src/, not inside it. This caused ENOENT errors when
trying to load extensions.

Fixed by calculating packageDir as the parent of currentDirPath
(i.e., the package root), then joining paths relative to that.