Ship pre-installed WordPress database in minified builds
The minified WordPress builds now include a pre-installed SQLite database,
saving clients the time required to run the WordPress installation wizard.

The database is included in wp.zip (which contains the PHP files that get
loaded into the virtual filesystem), but not in the static assets directory
that gets served separately.